Fix ac button and bs button
After clicking any of the digits if we press ac,
then any operand can be entered as operatorFlag
has become true, therefore turn the operatorFlag as
false when ac is pressed.

For issues related bs button: the match method returns
an array if match is found and null if no match is found. We
were trying to access null[0] which gives error and
the code execution stops never reaching the operate method