Fix decimal expression issue
After a digit is pressed, we can press decimal as
well but this can lead to a problem like multiple
decimals in a single operand like 1.234.56.78. So
using regex make sure if expression is of form 1.23
or 1.23 + 4.56, operation flag is false meaning
no decimal can be added to input expression.